+++ src/commands.ts
@@ -149,32 +149,44 @@ function calcDaysSinceIssued(issuedAt: number) {
   return diffDays;
 }
 
-export async function refreshAccessToken(context: vsc.ExtensionContext, licenseKey: string, accessToken?: string) {
+export function getDaysSinceIssued(accessToken?: string) {
+  const payload = accessToken != null ? jose.decodeJwt(accessToken) : null;
+  if (payload?.ent) return 0;
+  if (payload?.iat && accessToken) {
+    return calcDaysSinceIssued(payload.iat)
+  }
+  return Infinity;
+}
+
+function abortControllerTimeout(n: number) {
+  const ctrl = new AbortController();
+  setTimeout(() => ctrl.abort(), n);
+  return ctrl.signal;
+}
+
+export async function refreshAccessToken(context: vsc.ExtensionContext, daysSinceIssued: number, licenseKey: string, accessToken?: string) {
   let response;
   try {
     const baseURL = context.extensionMode === vsc.ExtensionMode.Development ? 'http://localhost:8788' : 'https://vscode.sqliteviewer.app';
-
-    const payload = accessToken != null ? jose.decodeJwt(accessToken) : null;
-    if (payload && 'ent' in payload) return accessToken;
-
-    const daysSinceIssued = accessToken && payload?.iat && calcDaysSinceIssued(payload.iat);
-    // console.log({ daysSinceIssued })
-    if (!daysSinceIssued || daysSinceIssued > 14) {
+    if (daysSinceIssued > 14) {
       response = await fetch(new URL('/api/register', baseURL), {
         method: 'POST',
         headers: [['Content-Type', 'application/x-www-form-urlencoded']],
         body: new URLSearchParams({ 'machine_id': await getShortMachineId(), 'license_key': licenseKey }),
+        signal: abortControllerTimeout(5000),
       });
-    } else if (daysSinceIssued > 1) {
+    } else if (daysSinceIssued > 1 && accessToken) {
       response = await fetch(new URL('/api/refresh', baseURL), {
         method: 'POST',
         headers: [['Content-Type', 'application/x-www-form-urlencoded']],
         body: new URLSearchParams({ 'machine_id': await getShortMachineId(), 'license_key': licenseKey, 'access_token': accessToken }),
+        // signal: abortControllerTimeout(5000),
       });
     } else {
       return accessToken;
     }
-  } catch {
+  } catch (err) {
+    if (err instanceof Error && err.name === 'AbortError') throw new Error('License validation request timed out (5000ms)');
     throw new Error('No response from license validation service');
   }